Refactor plugins tests
