--- Src/Plugins/Test.UnitTests.PlaintextSecrets/Test.UnitTests.PlaintextSecrets.csproj
@@ -1,10 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(SolutionDir)..\Targets\build.app.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.test.props" />
-
-  <ItemGroup>
-    <ProjectReference Include="..\PlaintextSecrets\PlaintextSecrets.csproj" />
-  </ItemGroup>
-
-</Project>
