--- Src/Plugins/Test.UnitTests.PotentiallySensitiveData/Test.UnitTests.PotentiallySensitiveData.csproj
@@ -1,10 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="$(SolutionDir)..\Targets\build.app.props" />
-  <Import Project="$(SolutionDir)..\Targets\build.test.props" />
-
-  <ItemGroup>
-    <ProjectReference Include="..\PotentiallySensitiveData\PotentiallySensitiveData.csproj" />
-  </ItemGroup>
-
-</Project>
