+++ Src/Plugins/Tests.BannedApi/BannedApiTests.cs
@@ -1,19 +1,23 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
+
+using Microsoft.CodeAnalysis.SarifPatternMatcher.Test.AllPlugins;
 
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Microsoft.CodeAnalysis.SarifPatternMatcher.Test.FunctionalTests.AllPlugins
+namespace Microsoft.CodeAnalysis.SarifPatternMatcher.Tests.BannedApi
 {
     public class BannedApiTests : EndToEndTests
     {
+        protected override string TestLogResourceNameRoot
+            => "Microsoft.CodeAnalysis.SarifPatternMatcher.Tests.BannedApi.TestData";
+
         public BannedApiTests(ITestOutputHelper outputHelper) : base(outputHelper) { }
 
         [Fact]
-        public void EndToEndFunctionalTests_BannedApi()
+        public void EndToEndFunctionalTests()
             => RunAllTests();
     }
 }
