+++ Src/Plugins/Tests.BannedApi/TestData/EmptyResults.sarif
