+++ Src/Plugins/Tests.BannedApi/TestData/ExpectedOutputs/Malloc.sarif
