+++ Src/Plugins/Tests.PlaintextSecrets/TestData/ExpectedOutputs/VariousPats.sarif
