+++ Src/Plugins/Tests.PlaintextSecrets/TestData/Inputs/VariousPats.txt
