+++ Src/Plugins/Tests.PotentiallySensitiveData/TestData/ExpectedOutputs/EmailAddresses.sarif
