+++ Src/Plugins/Tests.PotentiallySensitiveData/TestData/ExpectedOutputs/SocialSecurityNumbers.sarif
