+++ Src/Plugins/Tests.PotentiallySensitiveData/TestData/Inputs/EmailAddresses.txt
