+++ Src/Plugins/Tests.PotentiallySensitiveData/TestData/Inputs/SocialSecurityNumbers.txt
