+++ Src/Plugins/Tests.PotentiallySensitiveData/TestData/UnitTestOutput.PlaintextSecrets/ActualOutputs/VariousPats.sarif
