+++ Src/Plugins/Tests.PotentiallySensitiveData/TestData/UnitTestOutput.PlaintextSecrets/ExpectedOutputs/VariousPats.sarif
