+++ Src/Plugins/ValidationState.cs
@@ -1,11 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Microsoft.CodeAnalysis.SarifPatternMatcher.PlaintextSecrets
+namespace Microsoft.CodeAnalysis.SarifPatternMatcher.Plugins
 {
     public enum ValidationState
     {
